Fixed the handling of the "continuation" flag in SOUND commands
This means that "dummy" notes that are supposed to continue the previous pitch/envelope (channel number = &1**C instead of &0**C) no longer play the wrong note. I'm not sure where that wrong note was being picked up but I've simplified the code sequence and moved the check to just before the WriteCommand is issued and that seems to have fixed it.