Greatly simplified and standardised handling of colour/plot modes.
This means that each graphics mode doesn't need to handle setting colours/plot modes on its own, and instead just needs to pull the current colour and mode from PlotColour (which will be set to the logical 4-bit foreground or background colour automatically) and PlotMode (which will define the plotting mode, e.g. 0 for plot, 1 for OR, 2 for AND, 3 for EOR etc).

This has been implemented in MODE 2 (the 16-colour SMS mode) but has broken the 4-colour SMS MODE 1 (which assumes colour values are 2 bits and packed into a 4-bit value) and TMS9918 Graphics II MODE 3 (which was previously relying on COLOUR/GCOL to remap logical colours to TMS9918 colours and expects the foreground/background nybbles to be swapped).