Restored the drawing and colour support to the "Graphics II" mode.
All colour operations are now handled indirectly via a backup of the colour palette, allowing for better mapping of Master System colours to the TMS9918 palette.

For example, the TMS9918 dark blue is very dark, so is not mapped to "COLOUR 4" by default ("COLOUR 4" and "COLOUR 12" are the same).

However, if you really wanted that very dark blue, you could select it with "COLOUR 4,0,0,85").